requirements.txt
python-dotenv
telethon
PyYAML
pandas